# openapi/ziji-v1.yaml
...
openapi: 3.1.0
info:
  title: 资迹 Ziji API
  version: 0.1.0
  description: V1 正式 API 契约，对应 PRD V0.1、RTM V0.1 和《API 契约》V0.1；96 个 HTTP operationId 为正式实施基线。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: 当前主版本
security:
  - bearerAuth: []
...
paths:
...
  /auth/web/sessions:
    post:
      tags: [Auth]
      operationId: createWebSession
      summary: Web 邮箱密码登录
      security: []
      x-requirement-ids: [AUTH-002, AUTH-007]
      requestBody: { $ref: '#/components/requestBodies/LoginRequest' }
      responses:
        '201': { $ref: '#/components/responses/WebSessionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/mobile/sessions:
    post:
      tags: [Auth]
      operationId: createMobileSession
      summary: Mobile 邮箱密码登录
      security: []
      x-requirement-ids: [AUTH-002, AUTH-007]
      requestBody: { $ref: '#/components/requestBodies/LoginRequest' }
      responses:
        '201': { $ref: '#/components/responses/MobileSessionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/sessions/current:
    delete:
      tags: [Auth]
      operationId: revokeCurrentSession
      summary: 退出当前设备
      x-requirement-ids: [AUTH-004]
      responses:
        '204': { description: 当前设备会话已撤销 }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /users/me/password-change:
    post:
      tags: [Users]
      operationId: changePassword
      summary: 使用当前密码修改密码
      x-requirement-ids: [AUTH-003]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/PasswordChangeRequest' } }
      responses:
        '204': { description: 密码已修改 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { description: 未认证或会话失效返回 AUTHENTICATION_REQUIRED；当前密码错误、Hash 不支持或用户状态不允许改密返回 INVALID_CREDENTIALS；两者均不得泄露内部状态, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/me/sessions:
    get:
      tags: [Users]
      operationId: listUserSessions
      summary: 查询设备会话
      x-requirement-ids: [AUTH-004]
      parameters: [{ $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses:
        '200': { $ref: '#/components/responses/SessionListOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Users]
      operationId: revokeAllUserSessions
      summary: 撤销全部设备会话
      x-requirement-ids: [AUTH-004]
      responses:
        '204': { description: 全部设备会话已撤销 }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/me/sessions/{sessionId}:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    delete:
      tags: [Users]
      operationId: revokeUserSession
      summary: 撤销指定设备会话
      x-requirement-ids: [AUTH-004]
      responses:
        '204': { description: 设备会话已撤销 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webRefreshCookie:
      type: apiKey
      in: cookie
      name: ziji_refresh
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 500 }
...
    SessionId: { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
...
    LoginRequest:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } }
  schemas:
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [requestId, hasMore]
      properties:
        requestId: { type: string, minLength: 1, maxLength: 100 }
        nextCursor: { type: [string, 'null'], maxLength: 500 }
        hasMore: { type: boolean }
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [requestId]
      properties: { requestId: { type: string, minLength: 1, maxLength: 100 } }
    VersionConflictDetails:
      type: object
      additionalProperties: false
      required: [currentVersion, currentEtag, resourceLocation]
      properties:
        currentVersion: { type: integer, minimum: 1, description: 当前可见资源版本号 }
        currentEtag:
          type: string
          minLength: 3
          maxLength: 82
          pattern: '^"[1-9][0-9]*"$'
          description: 双引号包围的当前资源版本，例如 "7"。
        resourceLocation:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '^/(?!/)[^\s]*$'
          description: 以单个斜杠开头的相对 API 地址；不得以 // 开头。
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string, maxLength: 200 }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,79}$' }
        detail: { type: [string, 'null'], maxLength: 1000 }
        instance: { type: [string, 'null'], maxLength: 500 }
        requestId: { type: string, maxLength: 100 }
        versionConflict: { $ref: '#/components/schemas/VersionConflictDetails' }
        fieldErrors:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field: { type: string, maxLength: 200 }
              code: { type: string, maxLength: 80 }
              message: { type: [string, 'null'], maxLength: 500 }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, deviceName]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
        deviceName:
          type: string
          minLength: 1
          maxLength: 100
          description: 经 NFKC 与 trim 规范化的展示设备名称。
        deviceId:
          type: [string, 'null']
          maxLength: 200
          description: 可空客户端不透明稳定设备标识；非空时为 1～200 字符且不得全空白，保留客户端原值，不做 NFKC 或 trim。
...
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
        newPassword: { type: string, minLength: 10, maxLength: 128, writeOnly: true }
...
    Session:
      type: object
      additionalProperties: false
      required: [id, deviceName, createdAt, lastSeenAt, status]
      properties:
        id: { type: string, format: uuid }
        deviceName: { type: string, description: 经 NFKC 与 trim 规范化的 1～100 字符展示设备名称。 }
        deviceId: { type: [string, 'null'], description: NULL 或 1～200 字符的客户端不透明稳定设备标识；保留客户端原值，不做 NFKC/trim，且非空值不得全空白。 }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        status: { type: string, enum: [ACTIVE, REVOKED, EXPIRED] }
    TokenPair:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          minLength: 20
          description: RS256 access JWT；固定最长 30 分钟，不写入服务端持久化或日志。
        refreshToken:
          type: string
          minLength: 32
          description: 仅本次响应交付的轮换刷新凭据；安全存储后不得记录或回显。
        expiresIn: { type: integer, minimum: 1, description: Access Token 有效秒数，最长 30 分钟。 }
...
    WebSessionEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [session, accessToken, expiresIn]
          properties:
            session: { $ref: '#/components/schemas/Session' }
            accessToken:
              type: string
              minLength: 20
              description: RS256 access JWT；固定最长 30 分钟，不写入服务端持久化或日志。
            expiresIn: { type: integer, minimum: 1, description: Access Token 有效秒数，最长 30 分钟。 }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    MobileSessionEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [session, tokens]
          properties: { session: { $ref: '#/components/schemas/Session' }, tokens: { $ref: '#/components/schemas/TokenPair' } }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    SessionListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties: { data: { type: array, items: { $ref: '#/components/schemas/Session' } }, meta: { $ref: '#/components/schemas/PageMeta' } }
...
  responses:
    BadRequest: { description: 请求格式或字段校验失败, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Unauthenticated: { description: 未认证或会话失效, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    InvalidCredentials: { description: 邮箱不存在、密码错误、LOCKED 或 CLOSED 统一返回 INVALID_CREDENTIALS，不泄露账号存在、状态或密码 Hash 信息, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Forbidden: { description: 无对象权限或 CSRF 校验失败, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    NotFound: { description: 资源不存在或对当前用户不可见, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    RateLimited: { description: 请求过于频繁；Retry-After 为所有超限窗口中的最长剩余秒数, headers: { Retry-After: { description: 所有超限窗口中的最长剩余时间（秒）, schema: { type: integer, minimum: 1 } } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    WebSessionCreated: { description: "Web 会话已创建；响应以重复 Set-Cookie 设置 host-only 的 ziji_refresh（Secure、HttpOnly、SameSite=Strict、Path=/api/v1）与 ziji_csrf（Secure、非 HttpOnly、SameSite=Strict、Path=/api/v1），并设置 Cache-Control: no-store；刷新 Token 不出现在 JSON", content: { application/json: { schema: { $ref: '#/components/schemas/WebSessionEnvelope' } } } }
...
    MobileSessionCreated: { description: Mobile 会话与刷新 Token 已创建, content: { application/json: { schema: { $ref: '#/components/schemas/MobileSessionEnvelope' } } } }
...
    SessionListOk: { description: 设备会话列表, content: { application/json: { schema: { $ref: '#/components/schemas/SessionListEnvelope' } } } }
...